Move towards a more general Handler type
The split from Maiwar and HTTP means the handler type in the HTTP module
can be very specific. I always wanted the handler type higher up to be
much more general. This gets much of the way there. I still need to go
back over the static middleware and make it more general, but it's a
start.

I also really want to add an example which makes use of the simpler
handler, which doesn't stream at all. Parses all the input into some
type, serializes the output into a big ol' ByteString.