Try a different timeout method, also doesn't work
This has the advantage of not leaking a '<<timeout>>' exception, but the
draw back that it won't actually time out in lots of circumstances.
Including the two slowloris tests I've tried so far. So ... yeah. The
individual actions themselves need to be cancellable. I think this has
the advantage that (once I've remove use of Managed) each layer could be
cancelled with a more general 'timeout', say from unliftio.