Request timeouts
This is a bit of a raw stab. It actually still doesn't seem to close
connections on the slow loris read test, and I'm not sure why. I can see
the requests time out, but for whatever reason the connections remain
open.

I'd also like to just use MaybeT rather than Alternative (defaulting to
throwing mzero as I am).

    timeout :: MonadUnliftIO m => Int -> Pipe i o m a -> Pipe i o (MaybeT m) a
    timeout = _

I think I want the above in preference to what I have. So I need a
function that turns a `Stream o m (Maybe a)` into a
`Stream o (MaybeT m) a`. Feels doable.

But I also need to change the input stream at the same time. So a
`Stream i m a` into a `Stream i (MaybeT m) a`.